+++ csrank/objectranking/list_net.py
@@ -2,7 +2,6 @@ import logging
 
 from keras import backend as K
 from keras import Input
-from keras import optimizers
 from keras.layers import concatenate
 from keras.layers import Dense
 from keras.models import Model
@@ -34,7 +33,7 @@ class ListNet(Learner, ObjectRanker):
         kernel_regularizer=l2(1e-4),
         activation="selu",
         kernel_initializer="lecun_normal",
-        optimizer=SGD(),
+        optimizer=SGD,
         metrics=[zero_one_rank_loss_for_scores_ties],
         batch_size=256,
         random_state=None,
@@ -70,8 +69,10 @@ class ListNet(Learner, ObjectRanker):
                 Type of activation function to use in each hidden layer
             kernel_initializer : function or string
                 Initialization function for the weights of each hidden layer
-            optimizer : function or string
-                Optimizer to use during stochastic gradient descent
+            optimizer: Class
+                Uninitialized optimizer class following the keras optimizer interface.
+            optimizer__{kwarg}
+                Arguments to be passed to the optimizer on initialization, such as optimizer__lr.
             metrics : list
                 List of metrics to evaluate during training (can be
                 non-differentiable)
@@ -94,8 +95,7 @@ class ListNet(Learner, ObjectRanker):
         self.kernel_regularizer = kernel_regularizer
         self.kernel_initializer = kernel_initializer
         self.loss_function = loss_function
-        self.optimizer = optimizers.get(optimizer)
-        self._optimizer_config = self.optimizer.get_config()
+        self.optimizer = optimizer
         self.n_hidden = n_hidden
         self.n_units = n_units
         keys = list(kwargs.keys())
@@ -170,6 +170,7 @@ class ListNet(Learner, ObjectRanker):
         """
         self.random_state_ = check_random_state(self.random_state)
         _n_instances, _n_objects, self.n_object_features_fit_ = X.shape
+        self._initialize_optimizer()
         self._construct_layers(
             kernel_regularizer=self.kernel_regularizer,
             kernel_initializer=self.kernel_initializer,
@@ -213,7 +214,7 @@ class ListNet(Learner, ObjectRanker):
         merged = concatenate(outputs)
         model = Model(inputs=self.input_layer, outputs=merged)
         model.compile(
-            loss=self.loss_function, optimizer=self.optimizer, metrics=self.metrics
+            loss=self.loss_function, optimizer=self.optimizer_, metrics=self.metrics
         )
         return model
 
@@ -274,7 +275,7 @@ class ListNet(Learner, ObjectRanker):
             sess = tf.Session()
             K.set_session(sess)
             self._scoring_model = None
-            self.optimizer = self.optimizer.from_config(self._optimizer_config)
+            self._initialize_optimizer()
             self._construct_layers(
                 kernel_regularizer=self.kernel_regularizer,
                 kernel_initializer=self.kernel_initializer,
@@ -317,8 +318,8 @@ class ListNet(Learner, ObjectRanker):
         self.n_units = n_units
         self.kernel_regularizer = l2(reg_strength)
         self.batch_size = batch_size
-        self.optimizer = self.optimizer.from_config(self._optimizer_config)
-        K.set_value(self.optimizer.lr, learning_rate)
+        self._initialize_optimizer()
+        K.set_value(self.optimizer_.lr, learning_rate)
         self._construct_layers(
             kernel_regularizer=self.kernel_regularizer,
             kernel_initializer=self.kernel_initializer,